improve array info in getStructInfo.
members[x].isArray is true if the member is an array.
members[x].arrayCount is the array size, if it's an array.
members[x].isStruct is true if the member is a struct.
members[x].members are the members of the nested struct, if it's a struct.
A member can be an array of structs.